+++ Tests/SourceKittenFrameworkTests/DocInfoTests.swift
@@ -10,13 +10,19 @@ import Foundation
 import SourceKittenFramework
 import XCTest
 
+#if os(Linux)
+let sdkArgs: [String] = []
+#else
+let sdkArgs = ["-sdk", sdkPath()]
+#endif
+
 class DocInfoTests: XCTestCase {
 
     /// Validates that various doc string formats are parsed correctly.
     func testDocInfoRequest() {
         let swiftFile = File(path: fixturesDirectory + "DocInfo.swift")!
         let info = toNSDictionary(
-            try! Request.docInfo(text: swiftFile.contents, arguments: ["-sdk", sdkPath()]).send()
+            try! Request.docInfo(text: swiftFile.contents, arguments: sdkArgs).send()
         )
         compareJSONString(withFixtureNamed: "DocInfo", jsonString: toJSON(info))
     }
@@ -27,9 +33,8 @@ class DocInfoTests: XCTestCase {
             try! Request.moduleInfo(module: "",
                                     arguments: [
                                         "-c", swiftFile,
-                                        "-module-name", "DocInfo",
-                                        "-sdk", sdkPath()
-                ]).send()
+                                        "-module-name", "DocInfo"
+                ] + sdkArgs).send()
         )
         compareJSONString(withFixtureNamed: "ModuleInfo", jsonString: toJSON(info))
     }